requirements.txt
# Azure OpenAI Proxy Requirements

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Azure authentication
azure-identity==1.15.0

# OpenAI client
openai==1.3.0

# HTTP client
httpx==0.25.2

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment variable management
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0

# Retry mechanism
tenacity==8.2.3

# Optional: Performance monitoring
# gunicorn==21.2.0